Stop using util.Decode to decode header data

util.Decode calls binary.Read in a loop and each call allocates a byte slice.
Instead we can use binary.BigEndian methods directly.