@antoninbas 2025-01-08
Benchmark results.

Before:
```
$ go test -run=xxx -bench=BenchmarkCollectingProcess_DecodeDataRecord -benchmem ./pkg/collector/...
goos: darwin
goarch: amd64
pkg: github.com/vmware/go-ipfix/pkg/collector
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkCollectingProcess_DecodeDataRecord-16    	  270014	      3914 ns/op	     504 B/op	      21 allocs/op
PASS
ok  	github.com/vmware/go-ipfix/pkg/collector	17.955s
```

After:
```
$ go test -run=xxx -bench=BenchmarkCollectingProcess_DecodeDataRecord -benchmem ./pkg/collector/...
goos: darwin
goarch: amd64
pkg: github.com/vmware/go-ipfix/pkg/collector
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkCollectingProcess_DecodeDataRecord-16    	  349524	      3494 ns/op	     488 B/op	      14 allocs/op
PASS
ok  	github.com/vmware/go-ipfix/pkg/collector	22.516s
```

I was expecting a more significant improvement for CPU time, but at least the reduction in allocs/op is as expected (-7 as we decode 5 fields in the message header and 2 fields in the set header).